Fix cmd_duration in tmux
This is due to the iTerm shell integration doing funny things
We disable it for now.

In addition, we need to use a patched version of `starship init bash`
that doesn't try to be smart with not overwiriting previous DEBUG traps.
Essentially that means that we cannot use any DEBUG traps right now, but
we don't, so it's not that big of an issue.